refactor: move wage-garnishment to tools
