+++ apps/frontend/lib/tools/index.js
@@ -7,3 +7,4 @@ export {
   privateStudentLoanSchema,
 } from "./private-student-loan";
 export * from "./tax-offset";
+export * from "./wage-garnishment";
