+++ apps/frontend/lib/tools/wage-garnishment/index.js
@@ -1,0 +1,5 @@
+export { default as WageGarnishmentAForm, wageGarnishmentASchema } from "./a";
+export { default as WageGarnishmentBForm, wageGarnishmentBSchema } from "./b";
+export { default as WageGarnishmentCForm, wageGarnishmentCSchema } from "./c";
+export { default as WageGarnishmentDForm, wageGarnishmentDSchema } from "./d";
+export { default as WageGarnishmentEForm, wageGarnishmentESchema } from "./e";
